+++ database/pom.xml
@@ -102,6 +102,12 @@
       <version>${spring.version}</version>
     </dependency>
 
+    <dependency>
+      <groupId>org.springframework.data</groupId>
+      <artifactId>spring-data-jpa</artifactId>
+      <version>1.11.3.RELEASE</version>
+    </dependency>
+
   </dependencies>
 
 </project>
